feat(menu): extract to stack menu
