--- packages/picasso/src/FlatMenu/index.ts
@@ -1,6 +1,0 @@
-import { OmitInternalProps } from '@toptal/picasso-shared'
-
-import { Props } from './FlatMenu'
-
-export { default } from './FlatMenu'
-export type FlatMenuProps = OmitInternalProps<Props>
