--- packages/picasso/src/FlatMenuItem/index.ts
@@ -1,6 +1,0 @@
-import { OmitInternalProps } from '@toptal/picasso-shared'
-
-import { Props } from './FlatMenuItem'
-
-export { default } from './FlatMenuItem'
-export type FlatMenuItemProps = OmitInternalProps<Props>
