+++ packages/picasso/src/StackMenu/StackMenuContext.tsx
@@ -1,9 +1,9 @@
 import { ReactElement, createContext } from 'react'
 
-export interface FlatMenuContextProps {
+export interface StackMenuContextProps {
   push?: (key: string, menu: ReactElement) => void
   pop?: () => void
   refresh?: (key: string, menu: ReactElement) => void
 }
 
-export default createContext<FlatMenuContextProps>({})
+export default createContext<StackMenuContextProps>({})
