+++ packages/picasso/src/StackMenu/index.ts
@@ -1,0 +1,6 @@
+import { OmitInternalProps } from '@toptal/picasso-shared'
+
+import { Props } from './StackMenu'
+
+export { default } from './StackMenu'
+export type StackMenuProps = OmitInternalProps<Props>
