+++ packages/picasso/src/StackMenuItem/index.ts
@@ -1,0 +1,6 @@
+import { OmitInternalProps } from '@toptal/picasso-shared'
+
+import { Props } from './StackMenuItem'
+
+export { default } from './StackMenuItem'
+export type StackMenuItemProps = OmitInternalProps<Props>
